requirements.txt
# Основные зависимости
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.0.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-multipart==0.0.6

# Авторизация и безопасность
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Для работы с RAG
sentence-transformers==2.2.2
transformers==4.36.0
# chromadb==0.4.18
# chromadb-server
# chromadb-client[httpx]
# grpcio
# opentelemetry-exporter-otlp-proto-grpc
torch==2.1.0
numpy==1.24.3
accelerate==0.23.0
scikit-learn==1.3.2

# Для оценки качества RAG
ragas==0.2.0
datasets==2.14.5
pandas==2.1.1
seaborn==0.13.0
matplotlib==3.8.0
gradio==4.1.1

# Для обработки текста
beautifulsoup4==4.12.2
markdown==3.5
spacy==3.7.2
ru-core-news-md @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_md-3.7.0/ru_core_news_md-3.7.0-py3-none-any.whl

# Тестирование
pytest==7.4.3

# Если нужна поддержка GPU
# nvidia-cuda-runtime-cu118
# nvidia-cuda-nvrtc-cu118
# nvidia-cudnn-cu118
# bitsandbytes==0.41.1

ollama
PyJWT
qdrant-client
python-docx
PyPDF2
striprtf
odfpy
ebooklib
pyyaml
psycopg2-binary==2.9.9 